commit actividad2_4 9

+++ src/main/java/es/cifpcarlos3/Actividad2_4/model/Cuenta.java
@@ -13,7 +13,7 @@ public class Cuenta {
     private int idCuenta;
     private String numeroCuenta;
     private int idCliente;
-    private BigDecimal saldo;
+    private double saldo;
 
     @Override
     public String toString() {return "Cuenta: idCuenta: " + idCuenta
